f_op_actor_mng 99% (1 regswap left)

+++ include/d/d_item.h
@@ -393,7 +393,7 @@ BOOL isTriforce(u8 itemNo);
 BOOL isHeart(u8 itemNo);
 u8 getItemNoByLife(u8);
 u8 check_itemno(int);
-int getEmonoItemFromLifeBallTable(u16);
-int getItemFromLifeBallTableWithoutEmono(u16);
+u8 getEmonoItemFromLifeBallTable(u16);
+u8 getItemFromLifeBallTableWithoutEmono(u16);
 
 #endif /* D_ITEM_H */
